@@ -1,38 +1,38 @@
 module sigs.k8s.io/azurefile-csi-driver
 
-go 1.18
+go 1.20
 
 require (
-	github.com/Azure/azure-sdk-for-go v67.3.0+incompatible
+	github.com/Azure/azure-sdk-for-go v68.0.0+incompatible
 	github.com/Azure/azure-storage-file-go v0.8.0
 	github.com/Azure/go-autorest/autorest v0.11.28
-	github.com/Azure/go-autorest/autorest/adal v0.9.21 // indirect
+	github.com/Azure/go-autorest/autorest/adal v0.9.23 // indirect
 	github.com/Azure/go-autorest/autorest/to v0.4.0 // indirect
 	github.com/container-storage-interface/spec v1.7.0
 	github.com/gofrs/uuid v4.2.0+incompatible // indirect
 	github.com/golang/mock v1.6.0
-	github.com/golang/protobuf v1.5.2
+	github.com/golang/protobuf v1.5.3
 	github.com/kubernetes-csi/csi-lib-utils v0.7.0
 	github.com/kubernetes-csi/csi-proxy/client v1.0.1
 	github.com/kubernetes-csi/external-snapshotter/client/v4 v4.2.0
-	github.com/onsi/gomega v1.24.2
+	github.com/onsi/gomega v1.27.6
 	github.com/pborman/uuid v1.2.0
 	github.com/pelletier/go-toml v1.9.4
 	github.com/rubiojr/go-vhd v0.0.0-20200706105327-02e210299021
-	github.com/stretchr/testify v1.8.1
-	golang.org/x/net v0.5.0
-	google.golang.org/grpc v1.49.0
+	github.com/stretchr/testify v1.8.2
+	golang.org/x/net v0.9.0
+	google.golang.org/grpc v1.51.0
 	google.golang.org/protobuf v1.28.1
-	k8s.io/api v0.26.0
-	k8s.io/apimachinery v0.26.0
-	k8s.io/client-go v0.26.0
-	k8s.io/cloud-provider v0.26.0
-	k8s.io/component-base v0.26.0
-	k8s.io/klog/v2 v2.80.1
-	k8s.io/kubernetes v1.26.0
+	k8s.io/api v0.27.1
+	k8s.io/apimachinery v0.27.1
+	k8s.io/client-go v0.27.1
+	k8s.io/cloud-provider v0.27.1
+	k8s.io/component-base v0.27.1
+	k8s.io/klog/v2 v2.90.1
+	k8s.io/kubernetes v1.27.1
 	k8s.io/mount-utils v0.0.0
-	k8s.io/utils v0.0.0-20221128185143-99ec85e7a448
-	sigs.k8s.io/cloud-provider-azure v1.26.1-0.20230117121033-188ebdf2a417
+	k8s.io/utils v0.0.0-20230209194617-a36077c30491
+	sigs.k8s.io/cloud-provider-azure v1.27.1-0.20230424182702-d9d67257096a
 	sigs.k8s.io/yaml v1.3.0
 )
 
