@@ -19,14 +19,13 @@ package azurefile
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"reflect"
 	"runtime"
 	"strings"
 	"testing"
 
-	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2021-08-01/network"
+	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2022-07-01/network"
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
 
@@ -151,7 +150,7 @@ users:
 				}
 			}()
 
-			if err := ioutil.WriteFile(fakeKubeConfig, []byte(fakeContent), 0666); err != nil {
+			if err := os.WriteFile(fakeKubeConfig, []byte(fakeContent), 0666); err != nil {
 				t.Error(err)
 			}
 		}
@@ -173,7 +172,7 @@ users:
 			}
 			os.Setenv(DefaultAzureCredentialFileEnv, fakeCredFile)
 		}
-		cloud, err := getCloudProvider(test.kubeconfig, "", "", "", test.userAgent, test.allowEmptyCloudConfig, 5, 10)
+		cloud, err := getCloudProvider(test.kubeconfig, "", "", "", test.userAgent, test.allowEmptyCloudConfig, false, 5, 10)
 		if !testutil.AssertError(err, &test.expectedErr) && !strings.Contains(err.Error(), test.expectedErr.DefaultError.Error()) {
 			t.Errorf("desc: %s,\n input: %q, getCloudProvider err: %v, expectedErr: %v", test.desc, test.kubeconfig, err, test.expectedErr)
 		}
@@ -364,13 +363,17 @@ users:
 		}
 	}()
 
-	if err := ioutil.WriteFile(validKubeConfig, []byte(fakeContent), 0666); err != nil {
+	if err := os.WriteFile(validKubeConfig, []byte(fakeContent), 0666); err != nil {
 		t.Error(err)
 	}
+
+	os.Setenv("CONTAINER_SANDBOX_MOUNT_POINT", "C:\\var\\lib\\kubelet\\pods\\12345678-1234-1234-1234-123456789012")
+	defer os.Unsetenv("CONTAINER_SANDBOX_MOUNT_POINT")
 
 	tests := []struct {
 		desc                     string
 		kubeconfig               string
+		enableWindowsHostProcess bool
 		expectError              bool
 		envVariableHasConfig     bool
 		envVariableConfigIsValid bool
@@ -378,6 +381,7 @@ users:
 		{
 			desc:                     "[success] valid kube config passed",
 			kubeconfig:               validKubeConfig,
+			enableWindowsHostProcess: false,
 			expectError:              false,
 			envVariableHasConfig:     false,
 			envVariableConfigIsValid: false,
@@ -385,6 +389,15 @@ users:
 		{
 			desc:                     "[failure] invalid kube config passed",
 			kubeconfig:               emptyKubeConfig,
+			enableWindowsHostProcess: false,
+			expectError:              true,
+			envVariableHasConfig:     false,
+			envVariableConfigIsValid: false,
+		},
+		{
+			desc:                     "[failure] empty Kubeconfig under container sandbox mount path",
+			kubeconfig:               "",
+			enableWindowsHostProcess: true,
 			expectError:              true,
 			envVariableHasConfig:     false,
 			envVariableConfigIsValid: false,
@@ -392,7 +405,7 @@ users:
 	}
 
 	for _, test := range tests {
-		_, err := getKubeConfig(test.kubeconfig)
+		_, err := getKubeConfig(test.kubeconfig, test.enableWindowsHostProcess)
 		receiveError := (err != nil)
 		if test.expectError != receiveError {
 			t.Errorf("desc: %s,\n input: %q, GetCloudProvider err: %v, expectErr: %v", test.desc, test.kubeconfig, err, test.expectError)
