@@ -28,7 +28,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2021-08-01/network"
+	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2022-07-01/network"
 	"sigs.k8s.io/cloud-provider-azure/pkg/azureclients/subnetclient/mocksubnetclient"
 	azcache "sigs.k8s.io/cloud-provider-azure/pkg/cache"
 
@@ -542,8 +542,9 @@ func TestCreateVolume(t *testing.T) {
 			name: "storeAccountKey must set as true in cross subscription",
 			testFunc: func(t *testing.T) {
 				allParam := map[string]string{
-					subscriptionIDField:  "abc",
-					storeAccountKeyField: "false",
+					subscriptionIDField:              "abc",
+					storeAccountKeyField:             "false",
+					selectRandomMatchingAccountField: "true",
 				}
 
 				req := &csi.CreateVolumeRequest{
@@ -564,6 +565,37 @@ func TestCreateVolume(t *testing.T) {
 					})
 
 				expectedErr := status.Errorf(codes.InvalidArgument, "resourceGroup must be provided in cross subscription(abc)")
+				_, err := d.CreateVolume(context.Background(), req)
+				if !reflect.DeepEqual(err, expectedErr) {
+					t.Errorf("Unexpected error: %v", err)
+				}
+			},
+		},
+		{
+			name: "invalid selectRandomMatchingAccount value",
+			testFunc: func(t *testing.T) {
+				allParam := map[string]string{
+					selectRandomMatchingAccountField: "invalid",
+				}
+
+				req := &csi.CreateVolumeRequest{
+					Name:               "random-vol-name-selectRandomMatchingAccount-invalid",
+					CapacityRange:      stdCapRange,
+					VolumeCapabilities: stdVolCap,
+					Parameters:         allParam,
+				}
+
+				d := NewFakeDriver()
+				d.cloud = &azure.Cloud{
+					Config: azure.Config{},
+				}
+
+				d.AddControllerServiceCapabilities(
+					[]csi.ControllerServiceCapability_RPC_Type{
+						csi.ControllerServiceCapability_RPC_CREATE_DELETE_VOLUME,
+					})
+
+				expectedErr := status.Errorf(codes.InvalidArgument, "invalid selectrandommatchingaccount: invalid in storage class")
 				_, err := d.CreateVolume(context.Background(), req)
 				if !reflect.DeepEqual(err, expectedErr) {
 					t.Errorf("Unexpected error: %v", err)
