@@ -20,20 +20,13 @@ limitations under the License.
 package mounter
 
 import (
-	"runtime"
-
 	mount "k8s.io/mount-utils"
 	utilexec "k8s.io/utils/exec"
 )
 
-func NewSafeMounter() (*mount.SafeFormatAndMount, error) {
-	mounter := mount.New("")
-	if runtime.GOOS == "linux" {
-		// MounterForceUnmounter is only implemented on Linux now
-		mounter = mounter.(mount.MounterForceUnmounter)
-	}
+func NewSafeMounter(enableWindowsHostProcess bool) (*mount.SafeFormatAndMount, error) {
 	return &mount.SafeFormatAndMount{
-		Interface: mounter,
+		Interface: mount.New(""),
 		Exec:      utilexec.New(),
 	}, nil
 }
